Increase RelativePopulation precision from int to double
Increase precision to facilitate enabling SSM on stable at a less then 1% enable rate.

Bug: 338233130
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6049428
Commit-Queue: Mathijs Affourtit <maffourtit@google.com>
Cr-Commit-Position: refs/heads/main@{#1391829}

+++ chrome/common/profiler/thread_profiler_configuration.h
@@ -113,7 +113,7 @@ class ThreadProfilerConfiguration {
   // one of a set of variations.
   struct Variation {
     VariationGroup group;
-    int weight;
+    double weight;
   };
 
   ThreadProfilerConfiguration();
